Fix: mac address change during oq watchdog test (#20607)
Description of PR
Recently, a new QoS test case was added, which triggered an issue:
https://migsonic.atlassian.net/browse/MIGSMSFT-1203

The root cause is that during the test case, the system MAC address is changed after a soft reset. The script that

changes the MAC address was introduced in the following upstream PR: #8149

After debugging and discussion, it is confirmed that the script was added to fix a TD2-specific issue and is not applicable to Cisco devices.

Following the discussions, we implemented a fix to skip the MAC address setting for Cisco platforms.

signed-off-by: jianquanye@microsoft.com
Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>