Fix modules_changes test (#20322)
Description of PR
Summary: Update the kernel version used in the triggering command for the modules_changes test so it works across all HWSKUs and minor changes about fixture and checksum logic

Approach
What is the motivation for this PR?
Update the kernel version used in the triggering command for the modules_changes test so it works across all HWSKUs.
Fix checksum logic to depend on system bitness rather than Nokia/non-Nokia
Modify the fixture scope to minimize function calls.

Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>