[DHCPv6_Relay]: Ensuring Client Reachability (#19840)
What is the motivation for this PR?
It is observed in some cases "dhcp_relay/test_dhcpv6_relay.py::TestDhcpv6RelayWithMultipleVlan::test_dhcp_relay_default" fails if the neighbor entry for client LLA is not present.

How did you do it?
To make sure client address is resolved before we start the PTF test,
ping to a multicast address has been done so that we discover all the LLAs
in the segment.

How did you verify/test it?
Ran TestDhcpv6RelayWithMultipleVlan::test_dhcp_relay_default multiple times to make sure it passes.

Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>