use route creation for mellanox asic_type in test_crm_nexthop (#20320)
Adjust the crm_nexthop testcase for Mellanox after the recent changes
Starting 202505 we have an optimization in our SAI where we will program the nexthop to the hardware only when it has any references. Dangling nexthops are not programmed and kept in cache and will be programmed on the first reference. Hence we need a route referencing this nexthop to accurately reflect in CRM

Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>