Skipping nokia platforms for gcu test replace fec as fec is not configurable for 400g ports (#20542)
Approach
What is the motivation for this PR?
To skip test case execution for not supported platforms.

How did you do it?
Added the platforms to be skipped under the existing entry in tests/common/plpugins/conditional_mark/tests_mark_conditions.yaml

How did you verify/test it?
Ran test case in unsupported platforms and verified it is skipped.
Ran test case in cs-testbed and verified it is running.

Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>