[SmartSwitch] Restore dash smartswitch vnet after protobuf updates and pr 19700 merge (#20998)
Description of PR
This change restores the functionality of the dash smartswitch vnet testing after the latest changes to the proto buffers and pr 19700
The issue was that the vnet smartswitch test was failing becase the dash parser is now using the same method as for private link but is missing some cases used when parsing config templates for vnet.
Also some of the configurations from templates need to be updated based on the latest dash proto updates.

Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>