Issue 21314: Remove --enable_sflow_feature from sflow tests (#21319)
Summary:
Currently, sflow/test_sflow.py accepts a flag --enable_sflow_feature which will start the sflow feature when it is passed in. This is not necessary; if the feature is compiled in and available, then the test should enable it automatically when run, and if it was enabled by the test, we should disable it when it is finished.

Note that the test calls config reload at the end of the test anyways, so in practice it's not necessary to disable the feature after the test completes.

* remove --enable_sflow_feature from options
* unconditionally start sflow feature at start of test
* disable sflow feature at end of test if it was not enabled before

Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>