Fixing vnet bgp test so that it works for all T0 topologies (#20503)
What is the motivation for this PR?
The test is intended to work for all T0 topologies. Currently due to hard coded neighbor IPs, it does not work for some T0 topologies.

How did you do it?
Added code to fetch the ips, asn and other values from config_db before running the test and used those to

How did you verify/test it?
Validated against a few more T0 topologies such as vms-kvm-t0-64-32 which uses a different set of T1 neighbor IPs.

Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>