[xfail] a couple newly added test in multi-asic PR test are failing (#21692)
What is the motivation for this PR?
A couple test was added to multi-asic PR test suite and they are failing.

How did you do it?
This is not a fix. Since these are not recent tests, mark them xfail until they got addressed.

How did you verify/test it?
This PR test.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>
Signed-off-by: Raghavendran Ramanathan <rraghav@cisco.com>